# mempool: use regular map with mutex for mempool transaction senders