@alexanderbez 2021-05-17
Thanks @cuonglm but this PR isn't necessary as I've already addressed this in v1 of the mempool.